feat (auth.service): retira authguard do cadastro-usuario

+++ src/app/security/auth.service.ts
@@ -1,5 +1,6 @@
 // auth.service.ts
 import { Injectable } from '@angular/core';
+import { Route, Router } from '@angular/router';
 import { MessageService } from 'primeng/api';
 
 @Injectable({
@@ -7,7 +8,7 @@ import { MessageService } from 'primeng/api';
 })
 export class AuthService {
 
-  constructor() { }
+  constructor(private router:Router) { }
 
   getToken(): string | null {
     return sessionStorage.getItem('token');
@@ -18,6 +19,10 @@ export class AuthService {
   }
 
   isLoggedIn(): boolean {
+    if(this.router.url !== '/cadastro-usuario') {
+      return true;
+    }
+    
     const token = this.getToken();
     if(token === null) {
         sessionStorage.clear()
